added a splash screen acitvity for first launch to force user permissions before starting main activity and location reporting
